refactor mock_cluster
