+++ new-mock-engine-store/src/lib.rs
@@ -4,10 +4,8 @@
 pub mod config;
 pub mod mock_cluster;
 pub mod mock_store;
-pub mod node;
-pub mod server;
-pub mod transport_simulate;
 
+pub use engine_store_ffi::ffi::interfaces_ffi;
 pub use mock_cluster::*;
 pub use mock_store::*;
 pub use tikv_util::{box_err, box_try, debug, error, info, warn};
