+++ new-mock-engine-store/src/mock_cluster/common.rs
@@ -1,0 +1,4 @@
+// Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
+
+pub use engine_store_ffi::TiFlashEngine;
+pub use proxy_server::config::ProxyConfig;
