+++ new-mock-engine-store/src/mock_cluster/server.rs
@@ -79,10 +79,10 @@ use tokio::runtime::Builder as TokioBuilder;
 use transport_simulate::SimulateTransport;
 use txn_types::TxnExtraScheduler;
 
-use super::*;
+use super::{common::*, *};
 use crate::{
     config::Config,
-    mock_cluster::{Cluster, ProxyConfig, Simulator, TestPdClient, TiFlashEngine},
+    mock_cluster::{Cluster, ProxyConfig, Simulator, TestPdClient},
 };
 
 type SimulateStoreTransport =
