+++ new-mock-engine-store/src/mock_store/mock_snapshot_impls.rs
@@ -2,7 +2,7 @@
 
 use std::sync::atomic::Ordering;
 
-use crate::{
+use super::{
     common::*,
     mock_core::*,
     mock_engine_store_server::{into_engine_store_server_wrap, write_to_db_data},
