+++ new-mock-engine-store/src/mock_store/mock_write_impls.rs
@@ -1,6 +1,6 @@
 // Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::{
+use super::{
     common::*,
     mock_core::*,
     mock_engine_store_server::{
